Create Main.py
This code is nearly operational.  But occasionally, the HPMA sensor gets in an oddball state and won't accept commands.  When this happens, it seems like we have to power-cycle the HMPA sensor to get it working again.    Other times, the HPMA is stuck in auto transmit and seems to ignore input requesting it to change state.